requirements.txt
albumentations
hydra-core
tqdm
fvcore
matplotlib
wandb
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.9.0+cu111
torchvision==0.10.0+cu111
torchaudio==0.9.0
